+++ common/util/BitUtils.h
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <optional>
-
+#include "common/util/assert.h"
 #include "common/util/Range.h"
+#include "common/common_types.h"
 
 constexpr int BITS_PER_BYTE = 8;
 template <typename T>
@@ -40,3 +41,26 @@ std::optional<Range<int>> get_bit_range(T value) {
 
   return Range<int>{start_bit, end_bit};
 }
+
+/*!
+ * Note: this only works on numbers that are greater than or equal to 0.
+ */
+template <typename T>
+std::optional<int> get_power_of_two(T in) {
+  if (in == 0) {
+    return std::nullopt;
+  }
+  if ((in & (in - 1)) == 0) {
+    for (int i = 0; i < 64; i++) {
+      if (in == (T(1) << i)) {
+        return i;
+      }
+    }
+    assert(false);
+  } else {
+    return std::nullopt;
+  }
+}
+
+bool integer_fits(s64 in, int size, bool is_signed);
+u32 float_as_u32(float x);